+++ src/textual/_import_app.py
@@ -16,6 +16,23 @@ class AppFail(Exception):
     pass
 
 
+def shebang_python(candidate: Path) -> bool:
+    """Does the given file look like it's run with Python?
+
+    Args:
+        candidate (Path): The candidate file to check.
+
+    Returns:
+        bool: ``True`` if it looks to #! python, ``False`` if not.
+    """
+    try:
+        with candidate.open("rb") as source:
+            first_line = source.readline()
+    except IOError:
+        return False
+    return first_line.startswith(b"#!") and b"python" in first_line
+
+
 def import_app(import_name: str) -> App:
     """Import an app from a path or import name.
 
@@ -35,9 +52,14 @@ def import_app(import_name: str) -> App:
     from textual.app import App, WINDOWS
 
     import_name, *argv = shlex.split(import_name, posix=not WINDOWS)
+    drive, import_name = os.path.splitdrive(import_name)
+
     lib, _colon, name = import_name.partition(":")
 
-    if lib.endswith(".py"):
+    if drive:
+        lib = os.path.join(drive, os.sep, lib)
+
+    if lib.endswith(".py") or shebang_python(Path(lib)):
         path = os.path.abspath(lib)
         sys.path.append(str(Path(path).parent))
         try:
@@ -62,7 +84,7 @@ def import_app(import_name: str) -> App:
                 except KeyError:
                     raise AppFail(f"App {name!r} not found in {lib!r}")
             else:
-                # Find a App class or instance that is *not* the base class
+                # Find an App class or instance that is *not* the base class
                 apps = [
                     value
                     for value in global_vars.values()
@@ -95,6 +117,8 @@ def import_app(import_name: str) -> App:
             app = getattr(module, name or "app")
         except AttributeError:
             raise AppFail(f"Unable to find {name!r} in {module!r}")
+
+        sys.argv[:] = [import_name, *argv]
 
     if inspect.isclass(app) and issubclass(app, App):
         app = app()